Duplicating connected accounts by throwing on on the logged in user filter; poor seperation of concerns but...
