+++ src/main/java/nz/gen/wellington/rsstotwitter/controllers/HomepageController.java
@@ -1,7 +1,6 @@
 package nz.gen.wellington.rsstotwitter.controllers;
 
 import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
 import nz.gen.wellington.rsstotwitter.model.*;
 import nz.gen.wellington.rsstotwitter.repositories.mongo.JobDAO;
 import nz.gen.wellington.rsstotwitter.repositories.mongo.TwitterHistoryDAO;
@@ -13,7 +12,6 @@ import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @Controller
@@ -23,7 +21,7 @@ public class HomepageController {
     private final JobDAO jobDAO;
     private final TwitterHistoryDAO twitterHistoryDAO;
 
-    private final List<Destination> allDestinations = Lists.newArrayList(Destination.MASTODON, Destination.TWITTER);
+    private final List<Destination> allDestinations = Lists.newArrayList(Destination.values());
 
     @Autowired
     public HomepageController(LoggedInUserFilter loggedInUserFilter, JobDAO jobDAO,
@@ -49,40 +47,13 @@ public class HomepageController {
 
 
             return new ModelAndView("feeds").
-                    addObject("accounts", connectedAccountsFor(loggedInUser)).
+                    addObject("accounts", loggedInUserFilter.connectedAccountsFor(loggedInUser)).
                     addObject("jobs", jobsWithActivity);
 
         } else {
             return new ModelAndView("homepage").
                     addObject("destinations", allDestinations);
         }
-    }
-
-    private List<ConnectedAccount> connectedAccountsFor(Account account) {
-        List<ConnectedAccount> accounts = Lists.newArrayList();
-        for (Destination destination : destinationsConnectedToAccount(account)) {
-            accounts.add(new ConnectedAccount(destination.getAccountUsername(account), destination, destination.getAccountUrl(account)));
-        }
-        return accounts;
-    }
-
-    private Set<Destination> destinationsConnectedToAccount(Account account) {
-        Set<Destination> connected = Sets.newHashSet();
-        if (isAccountConnectedToMastdon(account)) {
-            connected.add(Destination.MASTODON);
-        }
-        if (isAccountContentedToTwitter(account)) {
-            connected.add(Destination.TWITTER);
-        }
-        return connected;
-    }
-
-    private boolean isAccountConnectedToMastdon(Account account) {
-        return account.getMastodonAccessToken() != null;
-    }
-
-    private boolean isAccountContentedToTwitter(Account account) {
-        return account.getToken() != null && account.getTokenSecret() != null;
     }
 
 }
